How to convert string to an integer

+++ 0x05-pointers_arrays_strings/100-atoi.c
@@ -96,7 +96,7 @@ int puiss(int a, int b)
  */
 int _atoi(char *s)
 {
-	int val = 0, i = 0, sizeFirst = sizeFirstInt(s), j, k = 0, pos = 0, neg = 0;
+	long int val = 0, i = 0, sizeFirst = sizeFirstInt(s), j, k = 0, pos = 0, neg = 0;
 
 	while (i <= (int)strlen(s) - 1)
 	{
@@ -121,5 +121,5 @@ int _atoi(char *s)
 	}
 	if (neg > pos)
 		val *= (-1);
-	return (val);
+	return ((long int)val);
 }
